Update Database.sql
New names of Aruba Curacao BES and Bonaire, should be in use since 2011. Official code from ISO registration and World Stastics. (These are not the same as Airport/Lines use, as I used in the commit of a few days ago, but hope this commit will also correct that fauly insert CUR => CUW)

The addition in Defaulte has an error in the red numer in front of the Country name. Is this the sort order or is this a more important sequential number?