Update insert of Country-names and 3 code ISO
If this work is acceptable, an effort to include all geo_country_refference values in the List table under country will be continued. After that I will continue and might even find the way to use Github and local machine work combined.

Please let me know if I should continue  or I should not attempt to double the names of all Countries in two tables.