Rapid Keyboard/Shorthand entry system is improved and 97% complete. Need to add in non-strabismus Neuro fields to shorthand entry system.
May need to rethink how to initially present the shorthand method.  It is so easy to do but non-eyedocs may not have a clue what it does.   Eyedocs should find it to be refreshingly simple and fast. Other users’ shorthand would be nice to make this entry option match more clinicians’ shorthand, so it is super fast for more people’s brains, not just mine.  Think it and it is documented?  Not quite, but this shorthand method is way faster than defaults + point—click—type for every field or typing and tabbing around…

Neuro shorthand works for strabismus documentation, really quite efficient.

Clicking “Keyboard Entry” displays fields shorthand names. 

Shorthand Help pages 90% complete but lacking beauty. Not translation ready at all ;(  Need to write shorthand Neuro Help section.

Considering Global Shorthand input zone for doing shorthand entry for every section in one fell swoop, in one textarea. Yeah, why not? It would be hide-able field, with visibilty being a fluid preference like the rest of the form.  It is shown automatically if used and left open during the last provider encounter, or not shown if not used last encounter, but “showable” via a click…