Example of bringing in the jquery files, which is a special case.
Special case since the old jquery repositories do not contain
the actual js files (they need to be built), and this method also
only brings in the 1 needed file (the index.js is a bower limitation
and is not a big issue since th directory identifies it).

By migrating the current js packages to these standardized packages
will allow developers to more easily migrate up to a more current
version (and to 1 version) via standard grep and sed methods(and
with a huge amount of testing work, of course).