Spectacle Rx: Detailed
Added many of the details an optometrist and optician uses for sending Rx to Lab.  This is the first iteration for this.  DB fields need to be added still.  This will lead to an API to deliver glasses Rx to local (or Distant) optical shoppe/lab.  Will need to add frame data to complete this also.