One little evil apostrophe
and a few tabindexes.
These sections will be rewritten.
Currently there are a max of 4 current Rxs, and technically there should be room for more.  Then again, if there are more than 4, the doctor is most likely going to very depressed.  All the more reason to make it so the EMR can handle the problem, making the doctor happy again.  So it will be rewritten as a function.