Crazy Menu Work
I actually had to learn something about Knockout.js and I am a big fan.
This leaves out the decisions of what goes into the menu and where - it is just a representative example of what could be.
The end game per Kevin’s work is to link this to a DB driven Administrative option thing.
This extends Kevin’s menu a bit - menu items can include a fontawesome icon  or a shortcut instruction set if we need it.