fix php8 warnings (see comment for details)
More modern version of php throw a huge amount of php warnings in the
error log when use a variable that does not yet exists (it does not
break anything, but what happens is that in future versions of php this
will turn into an error at some point; so we try to avoid warnings to
prevent that in the future). In this case we use a shorthand function
that looks like:
$example ?? ''
This is basically the same as doing
if (isset($example)) {
  $example = $example;
} else {
  $example = '';
}
So, it will basically set $example to '' if the $example
variable does not yet exist and this then avoids the php warnings.